airline-assistant/requirements.txt
python_dotenv
openai
gradio
